Generalise sustain update logic
